Move from Leenode to Libera.Chat
The Freenode IRC network has been taken over by someone who managed to
somehow "buy" it from christel.  Contrary to his promises to not
interfere, he tried to gain control over the network, leading to the
original staff team leaving and creating a new IRC network.

While I'm not happy to see Freenode butchered like that, I'm happy to
find that Libera.Chat seems to be a sane and safe place for ratpoison
users to lurk.